+++ domain/src/community.rs
@@ -33,7 +33,7 @@ pub trait HasCommunityId {
 pub struct Community {
     /// Unique identifier for this community.
     pub id: CommunityId,
-    _luck: u16,
+    _luck: u8,
     /// Members belonging to this community, keyed by their ID.
     pub members: HashMap<MemberId, Member>,
     /// The event log position up to which this snapshot has been computed.
@@ -61,19 +61,19 @@ impl Community {
     }
 
     /// Overrides the luck score.
-    pub fn with_luck(mut self, luck: u16) -> Self {
+    pub fn with_luck(mut self, luck: u8) -> Self {
         self._luck = luck;
         self
     }
 
     /// Overrides the luck score using a normalised value in `[0.0, 1.0]`.
     ///
-    /// The value is scaled to the internal `u16` range and rounded. Note that
+    /// The value is scaled to the internal `u8` range and rounded. Note that
     /// `self.luck() == luck` is not guaranteed: `luck` may not be exactly
-    /// representable as a `u16`, so a round-trip through this setter and
+    /// representable as a `u8`, so a round-trip through this setter and
     /// [`luck`][Self::luck] may differ slightly.
     pub fn with_luck_f64(mut self, luck: f64) -> Self {
-        self._luck = (luck * u16::MAX as f64).round() as u16;
+        self._luck = (luck * u8::MAX as f64).round() as u8;
         self
     }
 
@@ -85,9 +85,9 @@ impl Community {
 
     /// Returns this community's luck score normalised to `[0.0, 1.0]`.
     ///
-    /// `0.0` is neutral luck; `1.0` is the maximum (`u16::MAX`).
+    /// `0.0` is neutral luck; `1.0` is the maximum (`u8::MAX`).
     pub fn luck(&self) -> f64 {
-        self._luck as f64 / u16::MAX as f64
+        self._luck as f64 / u8::MAX as f64
     }
 
     /// Adds `member` to the community. Returns `true` if the member was newly inserted,
