+++ domain/src/member.rs
@@ -23,7 +23,7 @@ pub struct Member {
     pub id: MemberId,
     /// The name shown to other members.
     pub display_name: String,
-    _luck: u16,
+    _luck: u8,
     /// The fruits currently held by this member.
     pub bag: Bag,
 }
@@ -52,27 +52,27 @@ impl Member {
     }
 
     /// Overrides the luck score.
-    pub fn with_luck(mut self, luck: u16) -> Self {
+    pub fn with_luck(mut self, luck: u8) -> Self {
         self._luck = luck;
         self
     }
 
     /// Overrides the luck score using a normalised value in `[0.0, 1.0]`.
     ///
-    /// The value is scaled to the internal `u16` range and rounded. Note that
+    /// The value is scaled to the internal `u8` range and rounded. Note that
     /// `self.luck() == luck` is not guaranteed: `luck` may not be exactly
-    /// representable as a `u16`, so a round-trip through this setter and
+    /// representable as a `u8`, so a round-trip through this setter and
     /// [`luck`][Self::luck] may differ slightly.
     pub fn with_luck_f64(mut self, luck: f64) -> Self {
-        self._luck = (luck * u16::MAX as f64).round() as u16;
+        self._luck = (luck * u8::MAX as f64).round() as u8;
         self
     }
 
     /// Returns this member's luck score normalised to `[0.0, 1.0]`.
     ///
-    /// `0.0` is neutral luck; `1.0` is the maximum (`u16::MAX`).
+    /// `0.0` is neutral luck; `1.0` is the maximum (`u8::MAX`).
     pub fn luck(&self) -> f64 {
-        self._luck as f64 / u16::MAX as f64
+        self._luck as f64 / u8::MAX as f64
     }
 
     /// Adds one instance of `fruit` to this member's bag.
